+++ internal/services/cdn/cdn_frontdoor_rule_resource_test.go
@@ -148,8 +148,8 @@ resource "azurerm_cdn_frontdoor_origin_group" "test" {
 resource "azurerm_cdn_frontdoor_origin" "test" {
   name                          = "accTestOrigin-%[1]d"
   cdn_frontdoor_origin_group_id = azurerm_cdn_frontdoor_origin_group.test.id
+  enabled                       = true
 
-  enabled                        = true
   certificate_name_check_enabled = false
   host_name                      = "contoso.com"
   http_port                      = 80