+++ internal/services/managedidentity/registration.go
@@ -1,14 +1,15 @@
-package msi
+package managedidentity
 
 import (
 	"github.com/hashicorp/terraform-provider-azurerm/internal/sdk"
 	"github.com/hashicorp/terraform-provider-azurerm/internal/tf/pluginsdk"
 )
 
-// TODO: we should probably rename this Identity, or move into Authorization
+type Registration struct {
+	autoRegistration
+}
 
-type Registration struct{}
-
+var _ sdk.TypedServiceRegistrationWithAGitHubLabel = Registration{}
 var _ sdk.UntypedServiceRegistrationWithAGitHubLabel = Registration{}
 
 func (r Registration) AssociatedGitHubLabel() string {
@@ -17,14 +18,22 @@ func (r Registration) AssociatedGitHubLabel() string {
 
 // Name is the name of this Service
 func (r Registration) Name() string {
-	return "Managed Service Identities"
+	return r.autoRegistration.Name()
+}
+
+func (r Registration) DataSources() []sdk.DataSource {
+	dataSources := []sdk.DataSource{}
+	dataSources = append(dataSources, r.autoRegistration.DataSources()...)
+	return dataSources
+}
+
+func (r Registration) Resources() []sdk.Resource {
+	return r.autoRegistration.Resources()
 }
 
 // WebsiteCategories returns a list of categories which can be used for the sidebar
 func (r Registration) WebsiteCategories() []string {
-	return []string{
-		"Authorization",
-	}
+	return r.autoRegistration.WebsiteCategories()
 }
 
 // SupportedDataSources returns the supported Data Sources supported by this Service
@@ -36,7 +45,5 @@ func (r Registration) SupportedDataSources() map[string]*pluginsdk.Resource {
 
 // SupportedResources returns the supported Resources supported by this Service
 func (r Registration) SupportedResources() map[string]*pluginsdk.Resource {
-	return map[string]*pluginsdk.Resource{
-		"azurerm_user_assigned_identity": resourceArmUserAssignedIdentity(),
-	}
+	return map[string]*pluginsdk.Resource{}
 }