+++ internal/services/managedidentity/user_assigned_identity_data_source_test.go
@@ -1,4 +1,4 @@
-package msi_test
+package managedidentity_test
 
 import (
 	"fmt"